Extract process.env variables to config/env
