+++ config/env.js
@@ -1,0 +1,8 @@
+import { DB_TYPES } from './dbTypes';
+
+export const HOST = process.env.HOSTNAME || 'localhost';
+export const PORT = process.env.PORT || '3000';
+export const ENV = process.env.NODE_ENV || 'development';
+
+export const DB_TYPE = process.env.DB_TYPE || DB_TYPES.MONGO;
+
