+++ server/config/express.js
@@ -9,7 +9,7 @@ import gzip from 'compression';
 import helmet from 'helmet';
 import unsupportedMessage from '../db/unsupportedMessage';
 import { sessionSecret } from '../../config/secrets';
-import { DB_TYPE, ENV } from '../../config/app';
+import { DB_TYPE, ENV } from '../../config/env';
 import { session as dbSession } from '../db';
 
 export default (app) => {
