+++ server/db/index.js
@@ -1,4 +1,4 @@
-import { DB_TYPE } from '../../config/app';
+import { DB_TYPE } from '../../config/env';
 import { DB_TYPES } from '../../config/dbTypes';
 
 let dbConfig = null;
